bdschism/pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "bdschism"
dynamic = ["version"]
description = "Tools for launching and evaluating Bay-Delta SCHISM runs"
license = { text = "MIT" }
authors = [
    { name = "California Department of Water Resources", email = "Eli.Ateljevich@water.ca.gov" }
]
readme = "README.md"
keywords = ["Bay-Delta", "SCHISM", "hydrodynamics"]

# This has to be coordinated with entries in .github/python-package-conda.yml and build_sphinx.yml
requires-python = ">=3.10,<3.13"

dependencies = [
    "schimpy>=1.12.1",
    "suxarray",
    "vtools>=3.9.6",
    "dms_datastore",
    "dynaconf",
    "click",
    "pyproj",
	"shapely>=2.0",
	"geopandas",
	"netCDF4",
	"osgeo",
	"rasterio",
    "pyhecdss"
]

[project.optional-dependencies]
doc = [
    "sphinx",
    "sphinx-rtd-theme",
	"sphinx-autodoc-typehints",
    "sphinx-click",
    "sphinx-design",
    "sphinx-argparse",
    "sphinxcontrib-bibtex",
    "sphinxcontrib-mermaid",
    "sphinxcontrib-xlsxtable",
    "myst-parser",
	"ghp-import",
	"ipython",
    "numpydoc",
    "pillow>=4.0.0",
    "nbsphinx",
    "docutils",
    "babel"	
]
test = [
    "pytest",
    "pytest-cov",
    "tox"
]

[tool.setuptools]
packages = ["bdschism"]
include-package-data = true

[tool.setuptools.package-data]
bdschism = ["config/*.yaml"]

[tool.setuptools_scm]
search_parent_directories = true
root = "../"

[tool.click]
# This section is not strictly necessary unless you are setting up Click-specific configurations.

[tool.flake8]
max-line-length = 100
ignore = ["E122", "E123", "E126", "E127", "E128", "E731", "E722"]
exclude = ["build", "bdschism/_version.py", "tests", "conda.recipe", ".git", "benchmarks", ".asv"]

[tool.pytest.ini_options]
norecursedirs = [".*", "*.egg*", "build", "dist", "conda.recipe"]
addopts = [
    "--junitxml=junit.xml",
    "--ignore=setup.py",
    "--ignore=run_test.py",
    "--tb=native",
    "--strict-markers",
    "--durations=20"
]
markers = [
    "serial: execute test serially (to avoid race conditions)"
]

[project.urls]
Homepage = "https://github.com/CADWRDeltaModeling/BayDeltaSCHISM"

[project.scripts]
bds = "bdschism.__main__:cli"
set_nudge = "bdschism.set_nudging:set_nudging_cli"
hot_from_hot = "bdschism.hotstart_from_hotstart:hotstart_newgrid_cli"
hot_date = "bdschism.hotstart_date:set_hotstart_date_cli"
hot_nudge_data = "bdschism.hotstart_nudging_data:hotstart_nudge_data_cli"
create_nudging = "schimpy.nudging:main"
uv3d = "bdschism.uv3d:interpolate_uv3d_cli"
uv3d_single = "bdschism.uv3d.single_uv3d:single_uv3d_cli"
ccf_gate_height = "bdschism.ccf_gate_height:ccf_gate_cli"
make_links = "bdschism.create_sflux_links:make_links"
calc_ndoi = "bdschism.calc_ndoi:calc_indoi_cli"
parse_cu = "bdschism.parse_cu:parse_cu_cli"
plot_bds_bc = "bdschism.plot_bds_bc:plot_bds_bc_cli"
analyze_dt = "bdschism.analyze_dt:analyze_dt_cli"
gen_elev2d = "bdschism.gen_elev2d:gen_elev2d_cli"
smscg_log_2_schism_th = "bdschism.smscg_log_2_schism_th:smscg_log_2_schism_th"
set_mode = "bdschism.set_mode:set_mode_cli"
cd_prep = "bdschism.channel_depletion:cd_prep_cli"
potw_prep = "bdschism.potw:potw_prep_cli"
csv_to_th = "bdschism.convert_csv_th:csv_to_th_cli"
postprocess_source_sink = "bdschism.source_sink_postprocess:postprocess_source_sink_cli"
source_sink_workflow = "bdschism.source_sink_workflow:source_sink_workflow_cli"
combine_hotstart = "bdschism.combine_hotstart:combine_hotstart_cli"
config = "bdschism.config_cli:config"
visit_list = "bdschism.generate_visit_filelist:visit_list_cli"
nc_metadata = "bdschism.nc_metadata:nc_metadata_cli"
combine_nc = "bdschism.combine_nc:combine_nc_cli"